stdext::sync::mutex

Trait MutexExt

pub trait MutexExt<T> {
    // Required method
    fn force_lock(&self) -> MutexGuard<'_, T>;
}

Extension trait with useful methods for std::sync::Mutex.

Required Methods

fn force_lock(&self) -> MutexGuard<'_, T>

Shorthand for mutex.lock().unwrap() with a better panic message.

This method is intended to be used in situations where poisoned locks are considered an exceptional situation and should always result in panic.

Examples

use std::sync::Mutex;
use stdext::prelude::*;

let lock = Mutex::new(1);

let n = lock.force_lock();
assert_eq!(*n, 1);